requirements.txt
pyarrow==15.0.2
requests>=2.31.0
pymysql>=1.1.0
python-dotenv>=1.0.1
